requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
async-timeout==4.0.3
aiofiles==23.2.1

apache-airflow==2.9.2
apache-airflow-providers-http==4.5.1
apache-airflow-providers-common-sql==1.8.0
apache-airflow-providers-common-io==1.4.0
apache-airflow-providers-fab==1.4.0

Flask==2.2.5
flask-graphql==2.0.1
flask-cors==3.0.10
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==2.5.1  # compatible with Airflow's SQLAlchemy
SQLAlchemy==1.4.54

graphene<3
graphql-core<3

pyspark==3.4.1
pandas==2.0.3  # compatible stable version
numpy==1.24.4  # compatible stable version

python-dotenv==1.0.1
requests==2.31.0  # known stable
urllib3==1.26.18  # known stable

dbt-core==1.7.8
dbt-postgres==1.7.8
psycopg2-binary==2.9.9

isort==5.13.2
black==24.4.2
pytest==8.2.1

Werkzeug==2.3.8
Jinja2==3.1.4

gunicorn==21.2.0

certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7

typing_extensions==4.11.0
python-dateutil==2.9.0.post0
pendulum==3.0.0
tzdata==2024.1
PyYAML==6.0.1

pipdeptree==2.13.1
